修改exit_group逻辑

+++ os/src/fs/pipe.rs
@@ -271,7 +271,7 @@ impl File for Pipe {
     fn get_perm(&self) -> crate::auth::PermStat {
         let stat = self.get_stat();
         let (mode, uid, gid) = (stat.mode, stat.uid, stat.gid);
-        let mode = FileMode::from_bits_truncate(mode);
+        let mode = FileMode::from_bits_truncate(mode as u16);
         PermStat { mode, uid, gid }
     }
